delete advanced feature update modal and reminder
