+++ public/_locales/en/messages.json
@@ -3282,21 +3282,6 @@
   "agreeAndContinue": {
     "message": "Agree & Continue"
   },
-  "templeUpdate": {
-    "message": "Temple Update"
-  },
-  "templeWalletUpdate": {
-    "message": "Temple Wallet Update"
-  },
-  "templeUpdateDescription": {
-    "message": "Our Terms of Use, Privacy Policy have updated and we’ve added new section in settings for easier advanced features control. Please review and agree to the new terms to continue. "
-  },
-  "templeWalletUpdateDescription": {
-    "message": "Our Terms of Use and Privacy Policy have been updated. Please review the new terms and agree to ensure a smooth experience with your wallet."
-  },
-  "reviewUpdate": {
-    "message": "Review update"
-  },
   "buyWithCrypto": {
     "message": "Buy with Crypto"
   },
