--- src/app/pages/Home/OtherComponents/Tokens/components/TermsOfUseUpdateOverlay/selectors.ts
@@ -1,4 +1,0 @@
-export enum TermsOfUseUpdateOverlaySelectors {
-  acceptTermsCheckbox = 'Terms of Use Update Overlay/Accept Terms Checkbox',
-  closeButton = 'Terms of Use Update Overlay/Close Button'
-}
