+++ src/app/store/settings/state.mock.ts
@@ -1,4 +1,3 @@
-import { MAX_SHOW_AGREEMENTS_COUNTER, RECENT_TERMS_VERSION } from 'lib/constants';
 import { IS_MISES_BROWSER } from 'lib/env';
 import { mockPersistedState } from 'lib/store';
 
@@ -11,10 +10,7 @@ export const mockSettingsState = mockPersistedState<SettingsState>({
   isConversionTracked: false,
   pendingReactivateAds: false,
   adsImpressionsLinked: false,
-  acceptedTermsVersion: IS_MISES_BROWSER ? RECENT_TERMS_VERSION : 0,
   referralLinksEnabled: IS_MISES_BROWSER,
-  showAgreementsCounter: IS_MISES_BROWSER ? MAX_SHOW_AGREEMENTS_COUNTER : 0,
-  shouldShowTermsOfUseUpdateOverlay: true,
   toastsContainerBottomShift: 0,
   isTestnetModeEnabled: false
 });
