+++ src/app/store/settings/state.ts
@@ -1,6 +1,5 @@
 import { nanoid } from '@reduxjs/toolkit';
 
-import { MAX_SHOW_AGREEMENTS_COUNTER, RECENT_TERMS_VERSION } from 'lib/constants';
 import { IS_MISES_BROWSER } from 'lib/env';
 
 export interface SettingsState {
@@ -10,12 +9,15 @@ export interface SettingsState {
   isConversionTracked: boolean;
   pendingReactivateAds: boolean;
   adsImpressionsLinked: boolean;
-  acceptedTermsVersion: number;
   referralLinksEnabled: boolean;
-  showAgreementsCounter: number;
-  shouldShowTermsOfUseUpdateOverlay: boolean;
   toastsContainerBottomShift: number;
   isTestnetModeEnabled: boolean;
+  /** @deprecated */
+  acceptedTermsVersion?: number;
+  /** @deprecated */
+  showAgreementsCounter?: number;
+  /** @deprecated */
+  shouldShowTermsOfUseUpdateOverlay?: boolean;
   /** @deprecated */
   balanceMode?: 'fiat' | 'gas';
 }
@@ -27,10 +29,7 @@ export const settingsInitialState: SettingsState = {
   isConversionTracked: false,
   pendingReactivateAds: false,
   adsImpressionsLinked: false,
-  acceptedTermsVersion: IS_MISES_BROWSER ? RECENT_TERMS_VERSION : 0,
   referralLinksEnabled: IS_MISES_BROWSER,
-  showAgreementsCounter: IS_MISES_BROWSER ? MAX_SHOW_AGREEMENTS_COUNTER : 0,
-  shouldShowTermsOfUseUpdateOverlay: true,
   toastsContainerBottomShift: 0,
   isTestnetModeEnabled: false
 };
